feat(ultimate-backuper): pm2 startup WIP

+++ src/cli/prompts/createBackupTask.ts
@@ -1,5 +1,7 @@
 import prompts from 'prompts';
 import fs from 'fs';
+import path from 'path';
+import os from 'os';
 import { getBackupListFile } from '../../utils';
 
 export const buildCreateBackupTask = () => ({
@@ -19,6 +21,9 @@ export const buildCreateBackupTask = () => ({
     ]).then(({ name, dir }) => {
       if (!name) return console.log(`invalid task name ${name}`);
       if (!dir) return console.log(`invalid task dir ${dir}`);
+
+      if (!fs.existsSync(path.join(os.homedir(), dir)))
+        return console.log(`dir ${dir} doesn't exists`);
 
       const content = fs.readFileSync(getBackupListFile());
       fs.writeFileSync(getBackupListFile(), `${content}${name},${dir};`.trim());
